requirements.txt
annotated-types==0.7.0
blinker==1.8.2
cachelib==0.13.0
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
docstring_parser==0.16
Flask==3.0.3
Flask-Session==0.4.0
google-api-core==2.19.1
google-auth==2.33.0
google-cloud-aiplatform==1.60.0
google-cloud-bigquery==3.25.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.5
google-cloud-storage==2.18.2
google-crc32c==1.5.0
google-resumable-media==2.7.2
googleapis-common-protos==1.63.2
grpc-google-iam-v1==0.13.1
grpcio==1.65.4
grpcio-status==1.62.3
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
langdetect==1.0.9
MarkupSafe==2.1.5
numpy==2.0.1
packaging==24.1
proto-plus==1.24.0
protobuf==4.25.4
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.8.2
pydantic_core==2.20.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
requests==2.32.3
rsa==4.9
setuptools==72.1.0
shapely==2.0.5
six==1.16.0
typing_extensions==4.12.2
urllib3==2.2.2
vertexai==1.60.0
Werkzeug==3.0.3
wheel==0.44.0
